clean up build issues after merge
